sdk-polyscopex-5.0.36/samples/devices/device/requirements.txt
pytest==6.0.2
jsonschema==4.17.3
Flask==2.3.2
Flask-RESTful==0.3.9
pyserial==3.5